requirements.txt
torch
sentence-transformers
datasets
huggingface_hub
faiss-cpu
pandas
numpy
scikit-learn
rapidfuzz
gradio
pydantic
python-dotenv
pyyaml
pyarrow
